backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
google-generativeai==0.4.0
yfinance==0.2.31
pymongo==4.6.1
streamlit==1.30.0
requests==2.31.0
